Add ScipyOdeSimulator multiprocessing capability
This PR uses the multiprocessing module to add parallel simulation
capability to the ScipyOdeSimulator. This only works when using
Cython as a compiler.

Multiprocessing is picky about what can be pickled and sent to
child processes, so the Cython code is sent as text for use with
inline(). Since inline() is called in `__init__()` with the same code,
it utilises cython's cache, and therefore the ODEs only have to be
compiled once. The code is a little bit duplicative, since we can't
pass instance methods through to multiprocessing's processes, and
thefore the `rhs()` and `jac_fn()` methods must be redefined in
each process.